refactor(role): split role_archive into state / import / export submodules
原 src-tauri/src/api/role_archive.rs (933 行) 拆为 4 文件：

- mod.rs: #[tauri::command] 入口 + ImportResult/ExportResult 响应结构 + 重导出 state 类型

- state.rs: RoleArchiveState + ImportTaskEntry + ImportingGuard/TaskRemoveGuard (RAII)

- import_pipeline.rs: do_import, write_temp_archive, prepare_import_source, locate_extracted_dir, sanitize_role_folder_name, copy_dir_recursive, find_role_id_by_folder, parse_format, parse_policy

- export_pipeline.rs: compress_role_to_temp, sanitize_file_name

lib.rs 与 api/mod.rs 的调用点不变；api::role_archive::xxx 仍直接经 mod.rs 重导出可见。校验：cargo check --lib + cargo test --lib 19/19 通过（含 utils::archive 17 项 + emotion classifier 2 项）。
